okta: package for oie flow with Okta
Create the okta package which handles all the logic for communicating
with the Okta API, exchanging tokens and extracing the SAML response to
be passed to a Provider.

It starts by making a PreAuthorize call which requires user-interaction.
From this request we acquire the values necessary for continuing the
device authorization process, exchange tokens for web SSO tokens and
make the request for the SAML assertion which is sent via the
GenerateCredentials method in the Provider interface.